fix(audiod): virtio-snd slot stride must match audiod's page-aligned grant
audiod grants scratch pages with page-aligned stride (4096 bytes, since
space_grant requires page alignment), but virtio-snd was reading with
period_bytes stride (2048 bytes). Every odd period read from unwritten
padding = silence. Pattern: data-silence-data-silence = 43Hz buzz =
robotic distortion. Affected all playback regardless of sample rate.

Fix: virtio-snd computes slot_stride = (period_bytes + 4095) & !4095,
matching audiod's page-aligned grant stride.

Also fixes resampler i32 overflow in interpolate(): |diff| × frac could
exceed i32 range on full-scale transitions (65535 × 65535 = 4.3B >
i32::MAX). Switched to i64 arithmetic. Only affected resampled paths
(48k→44.1k, 22k→44.1k), not passthrough.

+++ userspace/audiod/src/resample.rs
@@ -162,11 +162,12 @@ impl LinearResampler {
 }
 
 /// Linear interpolation between two i16 samples.
-/// Returns `cur + (next - cur) * frac / FRAC_ONE` with i32 arithmetic.
+/// Returns `cur + (next - cur) * frac / FRAC_ONE` with i64 arithmetic
+/// to avoid overflow when |diff| × frac exceeds i32 range.
 #[inline]
 fn interpolate(cur: i16, next: i16, frac: u32) -> i16 {
-    let diff = (next as i32) - (cur as i32);
-    let interpolated = (cur as i32) + ((diff * frac as i32) >> FRAC_BITS);
+    let diff = (next as i64) - (cur as i64);
+    let interpolated = (cur as i64) + ((diff * frac as i64) >> FRAC_BITS);
     interpolated.clamp(-32768, 32767) as i16
 }
